etopay_sdk
Module types

Types module for SDK

Modules

currencies

Business logic for config sdk module

error

Errors related to sdk types

networks

Network definition

newtypes

Newtypes used for sensitive data This module contains newtype wrappers for sensitive values. There are several reasons for this:

transactions

business logic for transaction sdk module

users

Business logic for user sdk module

viviswap

business logic for viviswap sdk module

Structs

ApiApplicationMetadata

Export some api_types for the bindings to reference

CaseDetailsResponse

Export some api_types for the bindings to reference

File

Export some api_types for the bindings to reference A file which can be safely sent across the HTTP API by encoding the file bytes as base64.

IdentityOfficialDocumentData

Export some api_types for the bindings to reference

IdentityPersonalDocumentData

Export some api_types for the bindings to reference

KycAmlaQuestion

Export some api_types for the bindings to reference

KycOpenDocument

Export some api_types for the bindings to reference

NewCaseIdResponse

Export some api_types for the bindings to reference Request for getting a new posident case id

Order

Export some api_types for the bindings to reference

OrderList

Export some api_types for the bindings to reference Orders list

PaymentDetail

Export some api_types for the bindings to reference

Enums

ApiTxStatus

Export some api_types for the bindings to reference The state of the transaction

OfficialDocumentType

Export some api_types for the bindings to reference