etopay_sdk::types
Module viviswap

business logic for viviswap sdk module

Structs

NewViviswapUser

Struct for new viviswap user

ViviswapAddressDetail

Viviswap iban detail

ViviswapDeposit

Viviswap deposit contract details

ViviswapDepositDetails

Viviswap deposit details for FIAT to Crypto Swap

ViviswapKycStatus

Viviswap kyc status

ViviswapPartiallyKycDetails

The viviswap partial KYC details consisting of details for general and personal KYC steps

ViviswapState

Viviswap local app state

ViviswapWithdrawal

The viviswap withdrawal contract information

ViviswapWithdrawalDetails

Viviswap withdrawal details for crypto to FIAT swap

Enums

ViviswapDetailUpdateStrategy

When a Viviswap detail is added, there can be different ways to handle the logic

ViviswapVerificationStatus

Viviswap user verification status

ViviswapVerificationStep

Viviswap user verification step