etopay_sdk

Module wallet_manager

Wallet manager This module contains the implementation of the StardustWallet trait and the WalletImpl struct. The StardustWallet trait defines the methods for creating, migrating, backing up, restoring, and deleting a wallet. The WalletImpl struct represents an instantiated wallet and holds the necessary state and configuration.

Structs

WalletBorrow

Represents borrowing a [WalletUser] instance with a lifetime connected to the wallet manager. This prevents wallets to be stored and used later by another user.

WalletManagerImpl

Implementation of WalletManager that uses the SSS schema to store and retrieve the mnemonic and create implementations of [WalletUser].

Traits

WalletManager

Creates a wallet and returns an instance to work upon