etopay_sdk

Module core

Main SDK module.

Re-exports

pub use config::Config;

Modules

config

Config module. Configuration for SDK

exchange

Exchange module.

postident

Postident module. This module includes functions for interacting with postident operations.

share

Share module. The share module provides functionality for dealing with shares of the user wallet mnemonic.

transaction

Transaction module.

user

User module. This module defines methods for interacting with user-related functionality, such as getting user state, creating a new user, deleting a user, and more.

viviswap

Viviswap module. The Sdk struct is responsible for handling the Viviswap-related functionality and acts as a bridge between the Viviswap API and the application.

wallet

Wallet module. This module provides methods for initializing, verifying, deleting, and creating wallets, as well as migrating wallets from mnemonic or backup, creating backups, and verifying PINs.

Structs

Sdk

Struct representing the SDK and its core components including configuration, user management, and storage options.